+++ lib/govuk_publishing_components/presenters/brand_helper.rb
@@ -1,0 +1,17 @@
+module GovukPublishingComponents
+  module Presenters
+    class BrandHelper
+      def initialize(brand)
+        @brand = brand if brand
+      end
+
+      def get_brand
+        "brand--#{@brand}" if @brand
+      end
+
+      def get_brand_element(attribute)
+        "brand__#{attribute}" if @brand
+      end
+    end
+  end
+end